chore: move chains around and fix types after rebase
