+++ src/providers/ZapInitProvider/ModularZaps/HyperwaveZap.ts
@@ -12,7 +12,7 @@ import {
   http,
   parseUnits,
 } from 'viem';
-import { hyperevm } from '../hyperwave';
+import { hyperevm } from 'src/const/chains/hyperwave';
 import { buildContractComposable } from '../utils';
 import { approve, transfer } from './DefaultZap';
 import { ZapDefinition, ZapExecutionContext, ZapInstruction } from './base';
